fix: avatar pictures
