+++ packages/react-ui/src/app/components/sidebar/sidebar-user.tsx
@@ -4,7 +4,6 @@ import { LogOut } from 'lucide-react';
 
 import { useEmbedding } from '@/components/embed-provider';
 import { useTelemetry } from '@/components/telemetry-provider';
-import { Avatar, AvatarFallback } from '@/components/ui/avatar';
 import { Button } from '@/components/ui/button';
 import { SidebarMenu, SidebarMenuItem } from '@/components/ui/sidebar-shadcn';
 import {
@@ -14,6 +13,8 @@ import {
 } from '@/components/ui/tooltip';
 import { userHooks } from '@/hooks/user-hooks';
 import { authenticationSession } from '@/lib/authentication-session';
+import Avatar from 'boring-avatars';
+import { UserAvatar } from '@/components/ui/user-avatar';
 
 export function SidebarUser() {
   const { embedState } = useEmbedding();
@@ -27,11 +28,7 @@ export function SidebarUser() {
     <SidebarMenu>
       <SidebarMenuItem className="flex items-center justify-between w-full">
         <div className="flex items-center">
-          <Avatar className="h-8 w-8 rounded-lg">
-            <AvatarFallback className="rounded-lg bg-gray-200">
-              {user.firstName.charAt(0)}
-            </AvatarFallback>
-          </Avatar>
+          <UserAvatar name={user.firstName + ' ' + user.lastName} email={user.email} size={28} />
           <div className="grid flex-1 text-left text-sm leading-tight ml-2">
             <span className="truncate font-semibold">{user.firstName}</span>
             <span className="truncate text-xs">{user.email}</span>
