+++ packages/react-ui/src/components/ui/user-avatar.tsx
@@ -1,65 +1,37 @@
-import { useQueryClient } from '@tanstack/react-query';
-import { t } from 'i18next';
-import { LogOut } from 'lucide-react';
+import Avatar from "boring-avatars";
+import { Tooltip, TooltipTrigger, TooltipContent } from "./tooltip";
+import { Button } from "./button";
 
-import { useEmbedding } from '@/components/embed-provider';
-import { useTelemetry } from '@/components/telemetry-provider';
-import { userHooks } from '@/hooks/user-hooks';
-import { authenticationSession } from '@/lib/authentication-session';
+type UserAvatarProps = {
+  name: string;
+  email: string;
+  size: number;
+  disableTooltip?: boolean;
+};
 
-import { Avatar, AvatarFallback } from './avatar';
-import { AvatarLetter } from './avatar-letter';
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuTrigger,
-  DropdownMenuLabel,
-  DropdownMenuItem,
-} from './dropdown-menu';
-import { TextWithIcon } from './text-with-icon';
-export function UserAvatar() {
-  const { reset } = useTelemetry();
-  const { embedState } = useEmbedding();
-  const { data: user } = userHooks.useCurrentUser();
-  const queryClient = useQueryClient();
-  if (!user || embedState.isEmbedded) {
-    return null;
+export function UserAvatar({ name, email, size, disableTooltip = false }: UserAvatarProps) {
+  const tooltip = `${name} (${email})`;
+  const avatarElement = (
+      <Avatar
+        name={email}
+        size={size}
+        colors={['#0a0310', '#49007e', '#ff005b', '#ff7d10', '#ffb238']}
+        variant="beam"
+      />
+  );
+
+  if (disableTooltip) {
+    return avatarElement;
   }
 
   return (
-    <DropdownMenu>
-      <DropdownMenuTrigger asChild>
-        <Avatar className="cursor-pointer">
-          <AvatarFallback>
-            <AvatarLetter
-              name={user.firstName + ' ' + user.lastName}
-              email={user.email}
-              disablePopup={true}
-            ></AvatarLetter>
-          </AvatarFallback>
-        </Avatar>
-      </DropdownMenuTrigger>
-      <DropdownMenuContent align="end" className="w-[220px]">
-        <DropdownMenuLabel>
-          <div className="flex">
-            <div className="flex-grow flex-shrink truncate">{user.email}</div>
-          </div>
-        </DropdownMenuLabel>
-        <DropdownMenuItem
-          onClick={() => {
-            userHooks.invalidateCurrentUser(queryClient);
-            authenticationSession.logOut();
-            reset();
-          }}
-          className="cursor-pointer"
-        >
-          <TextWithIcon
-            icon={<LogOut size={18} className="text-destructive" />}
-            text={<span className="text-destructive">{t('Logout')}</span>}
-            className="cursor-pointer"
-          />
-        </DropdownMenuItem>
-      </DropdownMenuContent>
-    </DropdownMenu>
+    <Tooltip>
+      <TooltipTrigger asChild>
+        {avatarElement}
+      </TooltipTrigger>
+      <TooltipContent side="bottom">
+        {tooltip}
+      </TooltipContent>
+    </Tooltip>
   );
 }
