+++ packages/react-ui/src/features/team/component/invitation-card.tsx
@@ -7,10 +7,10 @@ import { useAuthorization } from '@/hooks/authorization-hooks';
 import { Permission, UserInvitation } from '@activepieces/shared';
 
 import { ConfirmationDeleteDialog } from '../../../components/delete-dialog';
-import { Avatar, AvatarImage } from '../../../components/ui/avatar';
 import { Button } from '../../../components/ui/button';
 import { userInvitationApi } from '../lib/user-invitation';
 import { userInvitationsHooks } from '../lib/user-invitations-hooks';
+import { UserAvatar } from '@/components/ui/user-avatar';
 
 export function InvitationCard({ invitation }: { invitation: UserInvitation }) {
   const { refetch } = userInvitationsHooks.useInvitations();
@@ -28,14 +28,11 @@ export function InvitationCard({ invitation }: { invitation: UserInvitation }) {
       key={invitation.id}
     >
       <div className="flex items-center space-x-4">
-        <Avatar className="hidden size-9 sm:flex">
-          <AvatarImage src="/avatars/05.png" alt={t('Avatar')} />
-          <AvatarFallback className="justify-center items-center flex">
-            <span className="p-2">
-              {invitation.email.charAt(0).toLocaleUpperCase()}
-            </span>
-          </AvatarFallback>
-        </Avatar>
+        <UserAvatar
+          name={invitation.email}
+          email={invitation.email}
+          size={32}
+        ></UserAvatar>
         <div>
           <p className="text-sm font-medium leading-none">
             {invitation.email} ({invitation.projectRole?.name})
